move `LSPLoop::runLSP` to live with the rest of `LSPLoop`
